module_6/src/worker/requirements.txt
# PostgreSQL Database Adapter
psycopg2-binary==2.9.9

# Environment variable management
python-dotenv==1.2.1

# RabbitMQ client
pika==1.3.2

# Web Scraping Dependencies
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.32.3
